Reverse skills in skillGroup when rendering matrix
In the paginated evaluation view, a user should progress through the levels of a category.

The UI displays this progress, and we want to ensure that when a user goes from one level to the next, that the current skill marker jumps to the next skill, not the top of the next level.