Frontend task list improvements
This commit makes the loading logic more sane and uses the react router Link component for each list item.

The reason for using the Link component is that it causes the Task component to be unmounted when the route changes. An unmounted component means that it will have to mount when it is viewed next time, which will trigger the retrieve tasks action to be dispatched (inside componentDidMount).